chore: split out piece handler code
